feat: support id generator
