+++ rust/Cargo.toml
@@ -36,6 +36,7 @@ byteorder = "1"
 mac_address = "1.1.4"
 hex = "0.4.3"
 time = "0.3.19"
+once_cell = "1.9.0"
 
 [build-dependencies]
 tonic-build = "0.7"
