+++ rust/src/models/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod message;
+pub(crate) mod message_id;
